Separate all TANF variables into their own files
