+++ openfisca_us/variables/hhs/tanf/cash/initial_tanf_eligibility.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class initial_tanf_eligibility(Variable):
+    value_type = bool
+    entity = SPMUnit
+    definition_period = YEAR
+    label = "Initial Economic Eligibility for TANF"
+    documentation = "Whether the familiy meets the economic requirements for the Temporary Assistance for Needy Families program on application."
