+++ openfisca_us/variables/hhs/tanf/cash/tanf.py
@@ -1,0 +1,19 @@
+from openfisca_us.model_api import *
+
+
+class tanf(Variable):
+    value_type = float
+    entity = SPMUnit
+    definition_period = YEAR
+    label = "TANF"
+    documentation = (
+        "Amount of Temporary Assistance for Needy Families benefit received."
+    )
+    unit = USD
+
+    def formula(spm_unit, period, parameters):
+        # Obtain eligibility.
+        eligible = spm_unit("is_tanf_eligible", period)
+        # Obtain amount they would receive if they were eligible.
+        amount_if_eligible = spm_unit("tanf_amount_if_eligible", period)
+        return where(eligible, amount_if_eligible, 0)
